Fixing some issues detected by the drone.io
Drone.IO appointed a missing header for the strdup function and warned
about the missing ';' character in a few lines of grammar file. Also,
the header cstdlib has been forgetten during the implementation of
tests and a leak had been detected in this file too. Some warnings
about type conversion had emerged.

+++ src/parser/TableChain.hpp
@@ -28,7 +28,7 @@ class TableChain
 
 		void push()
 		{
-			marks.push_back(data.size());
+			marks.push_back((unsigned int) data.size());
 		}
 
 		void pop()
@@ -61,7 +61,7 @@ class TableChain
 		iter_t find(unsigned line, const char *lexema, bool recursive = false)
 		{
 			typedef typename vector<T*>::reverse_iterator riter_t;
-			unsigned last = data.size() - lastIndex();
+			long unsigned int last = data.size() - lastIndex();
 			riter_t rend = (recursive) ? data.rend() : data.rbegin() + last;
 
 			(void) line; (void) lexema;
@@ -79,10 +79,10 @@ class TableChain
 
 		/* All this is just for test proposes */
 		/* { */
-		unsigned n_tokens() const
+		size_t n_tokens() const
 		{ return data.size(); }
 
-		unsigned n_sections() const
+		size_t n_sections() const
 		{ return marks.size(); }
 		/* } */
